% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sen_data.R
\name{get_senate_data}
\alias{get_senate_data}
\title{Load SenateData into the R environment}
\usage{
get_senate_data(states = NULL, related_to = "", years = NULL)
}
\arguments{
\item{states}{Default is NULL. If left blank, returns all states. Takes a
string or vector of strings of state names.}

\item{related_to}{Default is "". Provide a string to search a variable's name,
short/long descriptions from the codebook, and its citation for non-exact matches
of a search term.}

\item{years}{Default is NULL. If left blank, returns all years. Input can be
a single year (e.g. 2001) or a two-year vector representing the first and last
years desired (e.g. \code{c(1989, 2001)}).}
}
\value{
A tibble containing SenateData. The tibble has columns corresponding
to variables in the dataset, and rows corresponding to observations that match
the filtering criteria specified by \code{states}, \code{related_to}, and \code{years}.
}
\description{
Load SenateData into the R environment
}
\examples{
get_senate_data()
get_senate_data(states = "Michigan", years = c(2000, 2008))
get_senate_data(related_to = "tax")
}
