\name{get_senate_version}
\alias{get_senate_version}
\title{Get your local version of SenateData}
\usage{
get_senate_version()
}
\value{
No return value, called for side effects.
}
\description{
\code{get_senate_version} prints the current version of SenateData available through the package. It takes no arguments.
}
