% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_congress_version.R
\name{get_senate_version}
\alias{get_senate_version}
\title{Get your local version of SenateData}
\usage{
get_senate_version()
}
\value{
No return value, called for side effects.
}
\description{
\code{get_senate_version} prints the current version of Senate available through the package. It takes no arguments.
}
